+++ src/Conquest/TPlatform.h
@@ -2138,7 +2138,7 @@ SINGLE Platform<SaveStruct,InitStruct>::TestHighlight (const RECT & rect)
 				_rect.top = center_y - radius;
 				_rect.bottom = center_y + radius;
 
-				RECT screenRect = { 0, 0, (LONG)SCREENRESX, (LONG)SCREENRESY };
+				RECT screenRect = { 0, 0, SCREENRESX, SCREENRESY };
 
 				if ((bVisible = RectIntersects(_rect, screenRect)) != 0)
 				{
@@ -2185,7 +2185,7 @@ SINGLE Platform<SaveStruct,InitStruct>::TestHighlight (const RECT & rect)
 				_rect.top = center_y - radius;
 				_rect.bottom = center_y + radius;
 
-				RECT screenRect = { 0, 0, (LONG)SCREENRESX, (LONG)SCREENRESY };
+				RECT screenRect = { 0, 0, SCREENRESX, SCREENRESY };
 
 				if ((bVisible = RectIntersects(_rect, screenRect)) != 0)
 				{
