+++ src/Conquest/jumpgate.cpp
@@ -1805,7 +1805,7 @@ SINGLE Jumpgate::TestHighlight (const RECT & rect)	// set bHighlight if possible
 				_rect.top = center_y - radius;
 				_rect.bottom = center_y + radius;
 
-				RECT screenRect = { 0, 0, (LONG)SCREENRESX, (LONG)SCREENRESY };
+				RECT screenRect = { 0, 0, SCREENRESX, SCREENRESY };
 
 				if ((bVisible = RectIntersects(_rect, screenRect)) != 0)
 				{
